Implement Z80 register allocation with B'/C' support and optimizations
Major improvements to cc2 code generator:

Register Allocation:
- Fixed register allocation to support B' and C' byte registers
- B' and C' work with A (8-bit primary) since exx doesn't swap A
- Two-pass allocation: words first (BC, IX), then bytes (B, C, B', C')
- Proper BC/B/C conflict handling (BC allocation blocks B/C)
- Doubles byte register capacity from 2 to 4 registers

Code Generation Optimizations:
- Global long inline: ld (addr),hl; exx; ld (addr+2),hl; exx (no call)
- Framefree tail call: jp framefree instead of call framefree; ret
- Saves 3 bytes per function with frame

Test Suite:
- Added tests/test_load_store.c: comprehensive load/store test
- 40+ test functions covering all size/location combinations
- 8/16/32-bit operations across global/argument/local storage
- Validates all three major optimizations

Files changed:
- emit.c: B'/C' register operations with exx wrapping
- codegen.c: Two-pass register allocation strategy
- parseast.c: AST parser improvements
- tests/Makefile: Added test_load_store.c

All tests passing: 21 self-hosting files, 144 test suite files

+++ declare.c
@@ -79,7 +79,7 @@ declare_internal(struct type **btp, unsigned char struct_elem)
 {
     struct name *nm, *arg;
     struct type *t, *prefix, *suffix, *rt;
-    int i;
+    unsigned long i;
     unsigned char is_typedef_name;
     unsigned char kr_style;
     struct type *param_type;
